pyproject.toml
[project]
name = "wine-app"
version = "0.1.0"
description = "A web app for serving wine information"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.114.0",
    "jinja2>=3.1.4",
    "minio>=7.2.8",
    "pre-commit>=3.8.0",
    "python-dotenv>=1.0.1",
    "retry>=0.9.2",
    "uvicorn>=0.30.6",
]

[tool.setuptools]
py_modules = ["app"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W", "C90"]
ignore = ["E501"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
multiline-quotes = "double"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
testpaths = ["tests"]
